refs #80, added view in order to validate data from API
The view is here in order to help the front to validate data. The
complex validation done through presets cannot be running on the front
side. The view return a 204 , emtyp content, when the data are correct,
else returns 400 with the errors dictionary taken from the django form
The viewi is plugged in the set of API view available for the front
view.