refs #105, Add the metaclass to load permissions from settings
The permissions to access the view are fetch from settings. There is
three level of access, if no settings is provided, the default
permissions does not allow any access of the view. If the specific
settings is filled (specific to the view), the view loads theses
permissions. There is a global permissions settings key to fill to apply
the permissions on all the view.