Add a method in order to get the formidable object
In ValdiationView, the method to get the formidable object was "static".
Just check the "pk" key in the kwargs dict of the view. But, if the
client need another method to fetche the formidable object, no way was
offer. The main idea is to provide a method overridable in order to have
customisable point for the client.