Fix setup.py install bug
It looks like the .egg file generated when build "N" is happening is not always wiped out at the end of this build. As a consequence, the build "N+1" can't start because there's a "zip error" similar to this trac issue. [#1]

As a matter of fact, the global python setup.py install is useless, since tox will install the package in each isolated virtualenv. So the solution to this issue was to override the dependency setup by installing the most recent version of tox and avoid the misbehaving command.

1 - https://trac.edgewall.org/ticket/7014