refs #140, add field_map update in the dynamic form generation
In order to avoid custom field factory for each use case (When only one
field change for example), add a new argument in the builder method.
`field_map` update the inner mapping of the formfield_factory in order
to override the original mapping. With this mechanism no need to rewrite
an entire field factory just to override one kind of field