Add a ValidateView that works with ContextForm JSON
ValidateView needs a Formidable object.

We had issues in test because, when we need to delete a setting, tests
require to be decorated with `@override_settings()` without any
argument.

> You can also simulate the absence of a setting by deleting it after
> settings have been overridden, like this:
> @override_settings()
> def test_something(self):
>     del settings.LOGIN_URL
>     ...
> — django.test.modify_settings [1]

For CircleCI, we needed to upgrade `virtualenv` package as well as `tox`
because tests were failing with an old version of the package.

Also, we've renamed tests_integration.py → test_integration.py (to
follow the same convention with other test files). The related
django-perf-rec file has been renamed as well.

References:
[1] https://docs.djangoproject.com/en/1.8/topics/testing/tools/#django.test.modify_settings