$ref ConditionTest - items don't have a predefined type
When you're defining a condition that says that "Field X should be
displayed if Field Y is checked", it results in the following "condition
test":

```
{
  "field_id": "x",
  "operator": "eq",
  "values": [
    true
  ]
}
```

the value `true` is not a string, it's a plain boolean. Thus, the schema
definition was already wrong from this perspective, because
django-formidable already accepts "non-string" values.

Still, removing the type coercition will open the way to other type of
"condition tests". On the radar, we have the future "display if the
value selected in this list is 'X'" -- which will be a string, I know.
But could also mean that we could make conditional displays based on
number fields, etc. "show this date of birth field if you're under 21".