Adding tests on the `formidable.yml` file definition
This YAML file defines the accepted methods and formats for the form builder. It appeared that this is incomplete or obsolete and requires to be tested. As soon as something changes in this form definition, we should reflect it using a test that shows the impact.
It is recommended to use the TDD method, by writing the test first and implement the change on your second move.

The present test files are reflecting the spec *as it is written now* and will probably evolve as we'll fix them regarding the Python code that is supposed to implement this spec.