Added tests on the validation type (comparison operator)
At the moment, we don't know if it's possible to test the real
"validation grid" for types (some comparison operators are only
available if the field type is a date field).

The goal of this patch is to make sure that at least we're not trying to
validate an unknown operator.