Make sure that the `ValidateView.form_valid` method return an actual empty body content
This content is the spec-compatible content to pass along with a 204 HTTP status response.

Before this hotfix, the dictionary passed along as the response content was serialized into a 2-character string to calculate the content-length, but this content was not returned to the client. Some browsers would experience it badly, namely IE11.