Upgrading dependency to Django Rest Framework up to 3.9.x
This has led to raise an issue in our "Perf Rec" tests, that were trying to create forms with the wrong TestClient.
The consequence of this mishandling of the "POST" method was that the Formidable object was created, but none of its "children" (fields, accesses, validations, conditions, etc).

Switching to the DRF APITestClient and passing the `format='json'` argument led to fix the form creation. As a consequence, the SQL queries executed now reflect the creation of the objects in the Database *and* their retrieval afterwards.